Solved Sherlock and Array
Watson gives Sherlock an array A of length N. Then he asks him to
determine if there exists an element in the array such that the sum of
the elements on its left is equal to the sum of the elements on its
right. If there are no elements to the left/right, then the sum is
considered to be zero.
Formally, find an i, such that, A1+A2+...+A(i-1) = A(i+1) + A(i+2)+..AN.