BFS: Shortest Reach
Given an undirected graph consisting of N nodes (labelled 1 to N) where a specific given node S represents the start position
and an edge between any two nodes is of length 6 units in the graph.

It is required to calculate the shortest distance from start position (Node S) to all of the other nodes in the graph.